Check for 'org.gradle.java' in additional to 'java'.
This will hopefully address #40, that under Gradle's plugins mechanism,
the Java plugin is not recognized by the Protobuf plugin.

This is inspired by [this discussion
thread](https://discuss.gradle.org/t/project-plugins-hasplugin-java-returns-false-when-the-new-plugin-mechanism-is-used/2388),
while is no way to test it until it's released on Gradle plugins portal.

Also bumped the Gradle version requirement to 2.3 as mentioned in that
thread.