Allow protos in test to import protos in main.
Resolves #58

In Java projects, an `extractInclude*Proto` task will include files
under the `compileClasspath` of the corresponding source set. For
example, `extractIncludeTestProto` task will include the
`compileClasspath` of the `test` source set, which has the protos from
`main` source set.

In Android projects, the classpath trick doesn't work. Instead, we make
`extractIncludeAndroidTestProto` task include the protos of the `main`
source set, as well as the dependencies of its configuration, which is
`compile`.