Stop defaulting to javanano for Android.
protobuf-lite has superceded protobuf-javanano as the recommended
protobuf implementation for Android.  The protobuf plugin has been using
javanano as the default output on Android project.  It will stop doing
so.

Because the code generator for lite is only available on the most recent
protobuf versions (3.0.0), and it's not built-in protoc, but instead
provided as a protoc plugin, there isn't a way to reliably infer the
artifact spec for the javalite plugin out of a protoc spec.  Therefore
we will not default to any output for the Android plugin, and user needs
to add the javalite plugin manually.

Bump the version to 0.8.0-SNAPSHOT, as this is a behavioral change.

Changes to tests:
 - Protobuf version to 3.0.0
 - gRPC version to 1.0.0-pre2
 - testProject no longer tests nano.
 - Added testProjectLite, which tests lite only, and is depended by
 testProjectAndroid.
 - testProjectAndroid now generates for lite instead of nano.
 - testProjectAndroid now uses Android plugin 2.0.0-beta1
 - Tests stop using JUnit's TemporaryFolder, because it deletes the
 temporary projects whether or not the test succeeds, making it very
 difficult to debug.  Now the temporary projects are put under
 "build/tests".

The Gradle version used to build the plugin is upgraded to 2.14.1, as
required by the upgraded version of Android plugin.