Stop using a deprecated API and run tests for multiple Gradle versions.
Resolves #98

Additionally:

Make tests to run with the minimum (2.12) and the maximum (3.0)
supported Gradle versions, except for Android tests which require 2.14.1
as required by the Android plugin used.

Fix a typo in copying in #87 (Xmx->Xms)

Run Android test separately to reduce the chance that Travis kills the
Gradle daemon.  Also redirect Android test output to the console,
because it runs slightly longer than 10 minutes, and Travis will kill
the test if it has seen no output for 10 minutes.