Proto generation spreading (#174)
Spread proto generation over multiple protoc calls on systems that limit command line length, e.g., Windows. Once the command length exceeds the OS limit, it calls with the current arguments and resets the builder.

This is related to #167, allowing this issue to be avoided when many proto files are present, until google/protobuf#274 is completed.