WIP: add Gradle 5.0 in tests.
WIP: Still trying to fix tests. There are multiple issues:

1. extracted-include-protos as input of GenerateProtoTask, when
doesn't exist, will fail the validation of GenerateProtoTask.  Tried
to always mkdir for them, but doesn't work.  (TODO: Currently
suppressed by removing the "input" setting.  Should be fixed).

2. MethodDefNotFoundError from Kotlin plugin calling into Gradle API.
This is supposedly fixed by upgrading the kotlin plugin version in the
build file of the protouf plugin.

3. Could not find lint-gradle:26.1.0 in Android projects.  Probably
missing repo definition.

4. AndroidProjectDetectionTest fails due to class path issues.

Thought: currently the whole classpath of the protobuf plugin as it's
built is injected into the test projects (through
plugin-classpath.txt).  The plugin is built with Gradle 5.0 and a
certain Android and kotlin plugin versions, but the tests can be run
in different versions.  They may not necessarily be compatible.
Additionally, the current test harness is done in a way that's hard to
track, run and debug.  The whole thing needs to be revisited.