Fix a bug in Android plugin 2.5+ support.
Support for Android plugin 2.5+ was added in #121. While it made
extractIncludeProtosTask to copy the protos to a different destination
with Android plugin 2.5+ than with older verisons, it didn't make
generateProtoTask to "include" the new location.  This causes Android
tests to fail with Gradle 5.0 and Android plugin 3.1.0.

The bug has remained undetected because test was never run with
Android 2.5+.