Use separate configuration for extracting included proto from dependencies (#366)
This change creates a configuration that mirrors compileClasspath for accessing compilation dependencies. It is used by extractIncludeProtos task to extract proto files from upstream dependencies. The compileClasspath configuration itself would not work for dependencies using java-library plugin, as the optimization made by java-library plugin to only give the consumer project classes directory for compilation.

This change also bumps Gradle API version to 5.6 as the implementation requires to use LibraryElements API, which was introduced in Gradle 5.6.