make Utils.getConfigName a precise replica of DefaultSourceSet.configurationNameOf
The code that creates the `compileProtoPath` configuration for
a source set needs to find the full names of the `compileOnly`
and `implementation` configurations based on the source set name.

It isn't possible to read those names directly from the `SourceSet`
object, because `AndroidSourceSet` is not a `SourceSet`. Therefore,
we use the `Utils.getConfigName` method, which subsequently needs
to replicate the behavior of `DefaultSourceSet.configurationNameOf`.

The reason is that when a source set name is unconventional (e.g.
`native-test`), Gradle creates conventionally named configurations
(e.g. `nativeTestCompileOnly`). We need to do the same transformation
on the source set name, otherwise we won't find the configurations.