Fix configuration caching for Android projects
Avoid serializing entire task state and  extract necessary information
by using Providers. This avoids accessing the project instance during
execution.

When finding the path for the ExecutableLocator, instead of relying
on Task.doFirst additional task property is added. During execution,
this property is used to compute the final location of the executable.
This is necessary in order to capture ExecutableLocators correctly, as
final value used to be set during execution (which is too late).