Avoid using project.copy when possible and fix serialization (#406)
Since Gradle 6.0 there is FileSystemOperations API which is compatible
with configuration caching. This PR introduces CopyActionFacade which
exposes the copy method, but implementation is using the new API whne
possible, and reverts to project.copy when Gradle version is lower than 6.0.

This PR also starts using Gradle 6.0 to build the plugin. It is needed in
order to use FileSystemOperations API.

In GenerateProtoTask, SourceSet is now not serialized as part of the
task state, and a provider capturing what is needed is created instead.

Integration test for the Java plugin is updated to Gradle 6.5-rc-1.