Fix ProtobufExtract for Gradle Configuration Cache
by using an ArchiveActionFacade in the same vein as CopyActionFacade
in order to remove usage of Project at execution time

ArchiveActionFacade uses the internal FileOperations type from Gradle.
FileOperations is internal but quite stable and this plugin is already
making use of other internals. Moreover, a public alternative will be
available in Gradle 6.6

Signed-off-by: Paul Merlin <paul@gradle.com>