Add Kotlin DSL extension function for configuring proto source directory set in Android builds (#433)
This change adds the missing Kotlin DSL extension for configuring the `proto` SourceDirectorySet in Android build. The exiting extension has the receiver type of SourceSet while it should be AndroidSourceSet in Android builds. This caused method not found error when using the `proto` function in Android builds. This change add a compileOnly dependency on android build tools. An integration test covering Android build with Kotlin DSL is included.