Ignore OS as an input to `GenerateProtoTask`
This change will remove the OS from the input properties of the
GenerateProtoTask when `protoc` or a plugin is defined as a resolvable
artifact. The expectation is that same versions of `protoc` or plugins
will produce same outputs regardless of the OS on which they are run.

This is done by creating marking `alternativePaths` as Internal for
backwards compatibility and creating two new properties:
`releaseArtifacts` and `snapshotArtifacts`. `releaseArtifacts` will
contain "$groupId:$artifact:$version" for each non-snapshot dependency,
effectively ignoring OS for these dependencies. `snapshotArtifacts` will
contain each snapshot artifact in a `FileCollection` since the
snapshots cannot be matched solely by their maven coordinates.

See #457.