Ignore OS as an input to `GenerateProtoTask`
This change will remove the OS from the input properties of the
GenerateProtoTask when `protoc` is defined as a resolvable artifact. The
expectation is that same versions of protoc will produce same outputs
regardless of the OS on which it was run.

This is done by removing the "protoc" key from the `alternativePaths`
property when present and not a SNAPSHOT and adding a protocArtifactGav
property as an input that only considers the group, artifact ID, and
version of the resolvable `protoc`.

See #457.