More eagerly apply plugin
Tries to push ProtobufPlugin initialization as early as possible,
especially before afterEvaluate. This has the goal of reducing the
ordering requirements during configuration, making the plugin less
brittle internally. The public API is unchanged.

To make it more clear why delays are needed, initialization dependencies
were attempted to be made obvious. The code was refactored to try to
loop over objects only once. For example, looping over SourceSets only
once and doing all per-SourceSet initialization within that one loop and
doing that loop as soon as the SourceSets can be looped over. This was
not attempted for IDE support, partly because it is hard to test and
partly because we should investigate alternative configuration methods.
In the same vein, isAndroidProject() is passed explicitly because it is
an initialization dependency.