Workaround mapped value before task completed error for excluded task
When using provider.map, Gradle has checks that can fail if the provider
is used too early. For example:
> Querying the mapped value of provider(interface java.util.Set) before
> task ':p1:compileScala' has completed is not supported

However, it seems those checks can also fire even if the incomplete task
is a properly registered dependency and only when accessed from an
action. This is possible when using -x to exclude a task. This deserves
a Gradle bug, but let's workaround it in the meantime.

Fixes #550