Add generated proto sources to source set
Background:
Generated code was not in sourceSets. This is a problem for the task configuration. For example, the jar task does not package generated proto class. Before this change generated proto sources was injected directly to
compile tasks

Changes:
Add generateProtoTask to java sourceSet.

Test plan:
Green pipelines, production code logic was not changed.