Remove dependency on Guava
We like Guava, but the plugin has little need for it. As a side-benefit
dropping it avoids systems-not-handling-classpaths-correctly issues
like #247. We wouldn't drop Guava just for that reason (as the ecosystem
should fix its bugs), but instead of just removing it from the API
surface, let's remove it entirely. Similarly, we could continue using it
in our tests, but we really don't need it.

Fixes #572