Emulate and deprecate generatedFilesBaseDir (#636)
Fixes #33 for users not changing generatedFilesBaseDir. Stop documenting
generatedFilesBaseDir since it produces poor results, but keep it
functional with its existing Copy semantics. Fixes #332

This deprecates the setter, but I don't know how to make this trigger a warning,
for either Groovy and Kotlin. Both seem to ignore it.

Turn off UnnecessaryObjectReferences because it isn't providing value.