Delay buildDir evaluation until needed
This causes the plugin to use the correct configuration when users
use a non-default buildDir.

Note that for IDEs we create directories in afterEvaluate{}. It is
possible for the user's own afterEvaluate{} to change buildDir. In that
case, the task will output to the 'correct' directory but the mkdirs for
IDEs can be the 'old' buildDir. But users shouldn't be using
afterEvaluate{}, so we don't care and there's no real fix anyway.

Fixes #674